6-pop_listint.c

+++ 0x13-more_singly_linked_lists/6-pop_listint.c
@@ -9,7 +9,7 @@
 int pop_listint(listint_t **head)
 {
 	listint_t *new_head;
-	size_t copy;
+	int copy;
 
 	if (!*head)
 		return (0);
